feat: adding context of api data
