+++ src/assets/components/TopicsNavs.jsx
@@ -1,8 +1,12 @@
 import { Link } from "react-router-dom";
 import iconCar from '../icons/car.svg'
 import '../styles/components/TopicsNavs.css'
+import { useState, useEffect, useContext } from "react";
+import { VehiclesContext } from '../context/vehicles'
 
 function TopicsNavs() {
+    
+    const { vehicles, getVehicles } = useContext(VehiclesContext)
 
     return (
